+++ Source/DynamicScene/CzmlDataSource.js
@@ -25,6 +25,7 @@ define([
         '../Core/RuntimeError',
         '../Core/Spherical',
         '../Core/TimeInterval',
+        '../Core/TimeIntervalCollection',
         '../Scene/HorizontalOrigin',
         '../Scene/LabelStyle',
         '../Scene/VerticalOrigin',
@@ -85,6 +86,7 @@ define([
         RuntimeError,
         Spherical,
         TimeInterval,
+        TimeIntervalCollection,
         HorizontalOrigin,
         LabelStyle,
         VerticalOrigin,
@@ -743,15 +745,32 @@ define([
     }
 
     function processAvailability(dynamicObject, packet, dynamicObjectCollection, sourceUri) {
-        var availability = packet.availability;
-        if (!defined(availability)) {
+        var interval;
+        var packetData = packet.availability;
+        if (!defined(packetData)) {
             return;
         }
 
-        var interval = TimeInterval.fromIso8601(availability);
-        if (defined(interval)) {
-            dynamicObject.availability = interval;
+        var intervals;
+        if (Array.isArray(packetData)) {
+            var length = packetData.length;
+            for (var i = 0; i < length; i++) {
+                interval = TimeInterval.fromIso8601(packetData[i]);
+                if (defined(interval)) {
+                    if (!defined(intervals)) {
+                        intervals = new TimeIntervalCollection();
+                    }
+                    intervals.addInterval(interval);
+                }
+            }
+        } else {
+            interval = TimeInterval.fromIso8601(packetData);
+            if (defined(interval)) {
+                intervals = new TimeIntervalCollection();
+                intervals.addInterval(interval);
+            }
         }
+        dynamicObject.availability = intervals;
     }
 
     function processBillboard(dynamicObject, packet, dynamicObjectCollection, sourceUri) {